+++ mist-core/src/main/java/edu/snu/mist/core/master/DefaultQueryIdGeneratorImpl.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.master;
 
 import edu.snu.mist.core.parameters.QueryIdPrefix;
 import org.apache.reef.tang.annotations.Parameter;
