+++ mist-core/src/main/java/edu/snu/mist/core/master/QueryIdGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.master;
 
 import org.apache.reef.tang.annotations.DefaultImplementation;
