+++ mist-core/src/main/java/edu/snu/mist/core/rpc/DefaultClientToMasterMessageImpl.java
@@ -18,7 +18,7 @@ package edu.snu.mist.core.rpc;
 import edu.snu.mist.core.master.ApplicationCodeManager;
 import edu.snu.mist.core.master.MasterSetupFinished;
 import edu.snu.mist.core.master.allocation.QueryAllocationManager;
-import edu.snu.mist.core.task.QueryIdGenerator;
+import edu.snu.mist.core.master.QueryIdGenerator;
 import edu.snu.mist.formats.avro.ClientToMasterMessage;
 import edu.snu.mist.formats.avro.JarUploadResult;
 import edu.snu.mist.formats.avro.QuerySubmitInfo;
