+++ mist-core/src/test/java/edu/snu/mist/core/master/QueryIdGeneratorTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.mist.core.task;
+package edu.snu.mist.core.master;
 
 import edu.snu.mist.core.parameters.QueryIdPrefix;
 import edu.snu.mist.formats.avro.AvroDag;
